L'API ora si autoautentica in ambiente di Sviluppo, ovvero genera un token mock e lo assegna ad ogni richiesta che riceve (sostituendolo a quello che trova, eventualmente).

+++ ArcoApi/Models/ViewQlikDomandaValore.cs
@@ -5,7 +5,7 @@ namespace ArcoApi.Models
 {
     public partial class ViewQlikDomandaValore
     {
-        public double? IdDomandaValore { get; set; }
+        public int? IdDomandaValore { get; set; }
         public string Acronimo { get; set; }
         public string Testo { get; set; }
         public int? IdGruppo { get; set; }
